+++ packages/react-native-sdk/src/hooks/index.ts
@@ -5,3 +5,5 @@ export * from './useLocalVideoStream';
 export * from './useCallKeep';
 export * from './useIncallManager';
 export * from './usePublishMediaStreams';
+export * from './usePermissionRequest';
+export * from './usePermissionNotification';