+++ packages/react-native-sdk/src/icons/index.tsx
@@ -16,3 +16,4 @@ export * from './Cross';
 export * from './ScreenShare';
 export * from './ArrowRight';
 export * from './Reaction';
+export * from './VideoDisabled';