+++ packages/react-sdk/src/core/components/index.ts
@@ -1,5 +1,6 @@
 export * from './Audio';
 export * from './ParticipantView';
+export * from './StreamCall';
 
 export { Video } from './Video';
 export type { BaseVideoProps, VideoProps } from './Video';